Refactor snapcraft.storeapi
Refactor the 3 classes in __init.py__ in the snapcraft.storeapi package
into 3 different modules and expose those modules in __init__.py
